Implemented the basic Pixmap class.
This class will store the pixel map state. Currently, we maintain a
single 1D array of 8-bit unsigned integers to store the pixel color
values, initialized to all white. There are methods for getting and
setting pixel values; a line-drawing function which implements
Bresenham's algorithm; and a memcpy function to allow for safe
immutable updates.