Defined reducer to keep track of app state.
The reducer currently contains a Pixmap instance; a boolean flag to
indicate whether the user is currently drawing; the last position of the
mouse on the canvas, if definable; and the currently selected tool.
Most of the actions on the reducer are defined in terms of mouse events,
with an additional operation for setting the current tool.