Updated pixmap state to store color indices.
Instead of storing the RGB values directly in the pixmap state, we now
just store "color indices," which map to the appropriate RGB
configurations in our palette. In essence, the pixmap state is greatly
simplified since it does not need to store all the RGB pixel values.
Instead, we feed these into the canvas image data directly when
rendering it.