Refactored to use main/preview pixmap layers.
The app state now keeps track of a layers object, which contains two
Pixmap instances: one for the main, actual canvas; and another for the
"preview" canvas. As the name suggests, this is used to store pixel
information that has not been committed to the main canvas yet; for
instance, when previewing output while using the line tool.

If there is any info stored in the preview layer, it is rendered on the
actual canvas. However, once the preview is no longer needed (e.g., when
releasing the mouse button) the information from the preview is combined
with that in the main layer. In particular, since the preview is
intended to sit "on top" of the main canvas, any nonzero pixel
information stored in the preview will overrwrite the corresponding
entries in the main canvas, keeping it otherwise intact. The preview
canvas is then cleared.

It should be noted that, while we have not explicitly stated it here,
any state mutations on Pixmap data are safely handled by first copying
the instance, by invoking the copy method that was previously
implemented for the class.