Implemented basic color and tool bars.
The color bar now contains the 16 colors of the standard palette. (The
dithered colors will be handled later on.) For the tool bar, we have a
pretty elementary setup to begin with, since we'll be focusing more on
adding further tools before making any significant design changes.

I should note that there are a fair number of changes in this commit, to
the point that I tried to see if I could break them down into smaller
commits. Unfortunately that seems very difficult to do at this point,
and would impede me from making further progress. So, well, apologies to
my future self and/or anyone else reading this; but hopefully the code
makes sense.