Implemented variable brush thickness for ellipse drawing.
Similar to the approach taken for line drawing, the ellipse midpoint
algorithm is now abstracted to a generator function, which yields points
as they are calculated.

Also, the fill function was modified to implement 4-connected seed fill,
as opposed to 8-connected seed fill, to resolve issues of leaking when
attempting to fill closed regions. The tradeoff is that lines won't
completely fill if pixel-thin (thickness = 1), but this is a plausible
recreation of the fill behavior in the original application.