Implemented filled ellipse tool.
A filled ellipse is created by (a) drawing the outline, and (b)
iterating through each y-coordinate of the ellipse and drawing a line
between the corresponding min and max x-coordinates. This approach seems
to be the most efficient as of now, though it may require further
modifications in the future.

It should be noted that, to account for the nuance of additional pixels
when drawing ellipses with higher thicknesses, there is a condition in
the loop such that a pixel is only set to the fill color if it is not
already set to the border color. This only makes sense if we can
guarantee that a pixel has the border color iff it is part of the
border. I'll argue that this holds because we are only actually drawing
on the preview canvas, and so the only border-color pixels are those
that actually form the ellipse border. If I turn out to be wrong about
this, I'll aim to fix it in a future commit.