Set up toolbar icons with temp error handling.
Some of the tool icons (remastered in Figma) have been included as SVG
assets, and are loaded in the Toolbar component via a predefined
mapping. Since some of the icons are still work-in-progress, the
rectangle SVG is used as a temporary placeholder. Also, note that there
are icons for tools which do not yet have functional implementations.